Add Undo/Redo mechanism (#6)
Add SqliteHistory and JsonHistory associate with the SqliteStore and JsonStore.

JsonStore have been introduced here to merge the common methods for mashumaro, msgspec and pydantic stores.

SqliteHistory uses triggers to store the history into tables of the database.

JsonHistory save the whole data into an sqlite database. The tables of the database are close of those introduced for SqliteHistory. The database is saved along the json file.